requirements.txt
# TrustCoin Bot Dependencies
# Core Telegram Bot Library
python-telegram-bot==20.7

# Environment Variables Management
python-dotenv==1.0.0

# HTTP requests and async support
requests==2.31.0
aiohttp==3.9.1

# Web framework for webhooks
Flask==3.0.0

# For production deployment
gunicorn==21.2.0